VPR-59 refactor(cms): extract ActivityRow to satisfy the fallow gate
- The rail's row markup (actions cluster, verb caption, hover reveal)
  pushed RecentActivity's template past the complexity threshold; the
  row is now its own component and the shared item types live in
  CMS/types
